fix: NativeImage serialization of <webview>.capturePage() result (#20825)
