+++ lib/browser/guest-view-manager.js
@@ -5,6 +5,7 @@ const { ipcMainInternal } = require('@electron/internal/browser/ipc-main-interna
 const ipcMainUtils = require('@electron/internal/browser/ipc-main-internal-utils')
 const parseFeaturesString = require('@electron/internal/common/parse-features-string')
 const { syncMethods, asyncMethods } = require('@electron/internal/common/web-view-methods')
+const { serialize } = require('@electron/internal/common/type-utils')
 
 // Doesn't exist in early initialization.
 let webViewManager = null
@@ -381,6 +382,12 @@ handleMessageSync('ELECTRON_GUEST_VIEW_MANAGER_CALL', function (event, guestInst
   }
 
   return guest[method](...args)
+})
+
+handleMessage('ELECTRON_GUEST_VIEW_MANAGER_CAPTURE_PAGE', async function (event, guestInstanceId, args) {
+  const guest = getGuestForWebContents(guestInstanceId, event.sender)
+
+  return serialize(await guest.capturePage(...args))
 })
 
 // Returns WebContents from its guest id hosted in given webContents.
