+++ lib/browser/remote/server.ts
@@ -4,7 +4,7 @@ import * as electron from 'electron'
 import { EventEmitter } from 'events'
 import objectsRegistry from './objects-registry'
 import { ipcMainInternal } from '../ipc-main-internal'
-import { isPromise, isSerializableObject } from '@electron/internal/common/remote/type-utils'
+import { isPromise, isSerializableObject } from '@electron/internal/common/type-utils'
 
 const v8Util = process.electronBinding('v8_util')
 const eventBinding = process.electronBinding('event')
