+++ lib/browser/rpc-server.js
@@ -11,7 +11,7 @@ const { crashReporterInit } = require('@electron/internal/browser/crash-reporter
 const { ipcMainInternal } = require('@electron/internal/browser/ipc-main-internal')
 const ipcMainUtils = require('@electron/internal/browser/ipc-main-internal-utils')
 const guestViewManager = require('@electron/internal/browser/guest-view-manager')
-const clipboardUtils = require('@electron/internal/common/clipboard-utils')
+const typeUtils = require('@electron/internal/common/type-utils')
 
 const emitCustomEvent = function (contents, eventName, ...args) {
   const event = eventBinding.createWithSender(contents)
@@ -62,7 +62,7 @@ ipcMainUtils.handleSync('ELECTRON_BROWSER_CLIPBOARD', function (event, method, .
     throw new Error(`Invalid method: ${method}`)
   }
 
-  return clipboardUtils.serialize(electron.clipboard[method](...clipboardUtils.deserialize(args)))
+  return typeUtils.serialize(electron.clipboard[method](...typeUtils.deserialize(args)))
 })
 
 if (features.isDesktopCapturerEnabled()) {
