+++ lib/common/api/clipboard.js
@@ -4,13 +4,13 @@ const clipboard = process.electronBinding('clipboard')
 
 if (process.type === 'renderer') {
   const ipcRendererUtils = require('@electron/internal/renderer/ipc-renderer-internal-utils')
-  const clipboardUtils = require('@electron/internal/common/clipboard-utils')
+  const typeUtils = require('@electron/internal/common/type-utils')
 
   const makeRemoteMethod = function (method) {
     return (...args) => {
-      args = clipboardUtils.serialize(args)
+      args = typeUtils.serialize(args)
       const result = ipcRendererUtils.invokeSync('ELECTRON_BROWSER_CLIPBOARD', method, ...args)
-      return clipboardUtils.deserialize(result)
+      return typeUtils.deserialize(result)
     }
   }
