+++ lib/renderer/api/remote.js
@@ -4,7 +4,7 @@ const v8Util = process.electronBinding('v8_util')
 const { hasSwitch } = process.electronBinding('command_line')
 
 const { CallbacksRegistry } = require('@electron/internal/renderer/remote/callbacks-registry')
-const { isPromise, isSerializableObject } = require('@electron/internal/common/remote/type-utils')
+const { isPromise, isSerializableObject } = require('@electron/internal/common/type-utils')
 const { ipcRendererInternal } = require('@electron/internal/renderer/ipc-renderer-internal')
 
 const callbacksRegistry = new CallbacksRegistry()
