+++ typings/internal-electron.d.ts
@@ -162,6 +162,7 @@ declare namespace ElectronInternal {
 
     // Created in web-view-impl
     public getWebContentsId(): number;
+    public capturePage(rect?: Electron.Rectangle): Promise<Electron.NativeImage>;
   }
 }
